rename several files in the webview
