+++ src/documentation/webview/Disposable.ts
